requirements.txt
beautifulsoup4==4.13.4
certifi==2025.6.15
charset-normalizer==3.4.2
fake-useragent==2.2.0
idna==3.10
python-dateutil==2.9.0.post0
requests==2.32.4
six==1.17.0
soupsieve==2.7
urllib3==2.5.0
